requirements.txt
streamlit>=1.32.0
langchain==0.3.7
langchain-community==0.3.7
langchain-groq>=0.1.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
